Create README.md
Me and my teammates did a survey sampling project and our research topic was about the happiness of Middle East Technical University students. The major goal of this study is to analyze the relationship between the level of happiness of university students in METU and their social relationships, monthly income, and academic success. While doing this project, we collected data from METU students and conducted our research using SPSS. You can review the data we collected on github. You can see the statistical tests and graphs we used and the questions we asked in the report.